Cleanup Yii2ServiceProvider
