+++ yii2-adapter/legacy/services/Categories.php
@@ -29,7 +29,7 @@ use CraftCms\Cms\Support\Facades\Structures;
 use CraftCms\Cms\Support\Str;
 use CraftCms\Cms\Twig\TemplateResolver;
 use CraftCms\Cms\View\TemplateMode;
-use CraftCms\Yii2Adapter\Yii2ServiceProvider;
+use CraftCms\Yii2Adapter\DeprecatedConcepts;
 use Illuminate\Database\Query\Builder;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Auth;
@@ -126,7 +126,7 @@ class Categories extends Component
     private function _groups(): MemoizableArray
     {
         if (!isset($this->_groups)) {
-            if (!Yii2ServiceProvider::supportsCategories()) {
+            if (!DeprecatedConcepts::supportsCategories()) {
                 $this->_groups = new MemoizableArray([]);
 
                 return $this->_groups;
