+++ yii2-adapter/src/Console/AddGlobalSetsSupportCommand.php
@@ -9,7 +9,7 @@ use CraftCms\Cms\Cms;
 use CraftCms\Cms\Console\CraftCommand;
 use CraftCms\Cms\Database\Migrator;
 use CraftCms\Cms\Database\Table;
-use CraftCms\Yii2Adapter\Yii2ServiceProvider;
+use CraftCms\Yii2Adapter\DeprecatedConcepts;
 use Illuminate\Console\Command;
 use Illuminate\Console\ConfirmableTrait;
 use Illuminate\Database\Schema\Blueprint;
@@ -17,8 +17,8 @@ use Illuminate\Support\Facades\Schema;
 
 final class AddGlobalSetsSupportCommand extends Command
 {
-    use CraftCommand;
     use ConfirmableTrait;
+    use CraftCommand;
 
     protected $signature = 'craft:add-global-sets-support
         {--force : Force the operation to run when in production or when allowAdminChanges is disabled.}
@@ -33,6 +33,7 @@ final class AddGlobalSetsSupportCommand extends Command
 
         if (Schema::hasTable($globalSetsTable)) {
             $this->info("<fg=cyan>$globalSetsTable</> table already exists.");
+
             return;
         }
 
@@ -70,7 +71,7 @@ final class AddGlobalSetsSupportCommand extends Command
             },
         );
 
-        Yii2ServiceProvider::resetSupport();
+        DeprecatedConcepts::resetSupport();
     }
 
     protected function getDefaultConfirmCallback(): Closure
