+++ yii2-adapter/src/Console/DropGlobalSetsSupportCommand.php
@@ -8,15 +8,15 @@ use craft\db\Table as LegacyTable;
 use CraftCms\Cms\Cms;
 use CraftCms\Cms\Console\CraftCommand;
 use CraftCms\Cms\Database\Migrator;
-use CraftCms\Yii2Adapter\Yii2ServiceProvider;
+use CraftCms\Yii2Adapter\DeprecatedConcepts;
 use Illuminate\Console\Command;
 use Illuminate\Console\ConfirmableTrait;
 use Illuminate\Support\Facades\Schema;
 
 final class DropGlobalSetsSupportCommand extends Command
 {
-    use CraftCommand;
     use ConfirmableTrait;
+    use CraftCommand;
 
     protected $signature = 'craft:drop-global-sets-support
         {--force : Force the operation to run when in production or when allowAdminChanges is disabled.}
@@ -31,6 +31,7 @@ final class DropGlobalSetsSupportCommand extends Command
 
         if (!Schema::hasTable($globalSetsTable)) {
             $this->info("<fg=cyan>$globalSetsTable</> table doesn’t exist.");
+
             return;
         }
 
@@ -45,7 +46,7 @@ final class DropGlobalSetsSupportCommand extends Command
             },
         );
 
-        Yii2ServiceProvider::resetSupport();
+        DeprecatedConcepts::resetSupport();
     }
 
     protected function getDefaultConfirmCallback(): Closure
