+++ yii2-adapter/src/Console/MigrateSessionsTableCommand.php
@@ -13,8 +13,8 @@ use Illuminate\Support\Facades\Schema;
 
 final class MigrateSessionsTableCommand extends Command
 {
-    use CraftCommand;
     use ConfirmableTrait;
+    use CraftCommand;
 
     protected $signature = 'craft:migrate:sessions-table
         {--force : Force the operation to run when in production or when allowAdminChanges is disabled.}
