+++ yii2-adapter/src/Mixins/ValidateMixin.php
@@ -5,6 +5,12 @@ declare(strict_types=1);
 namespace CraftCms\Yii2Adapter\Mixins;
 
 use Closure;
+use CraftCms\Cms\Asset\Data\Volume;
+use CraftCms\Cms\Dashboard\Widgets\Widget;
+use CraftCms\Cms\Element\Element;
+use CraftCms\Cms\Field\Field;
+use CraftCms\Cms\FieldLayout\FieldLayoutComponent;
+use CraftCms\Cms\Filesystem\Filesystems\Filesystem;
 use CraftCms\Cms\Support\Arr;
 use CraftCms\Cms\Support\Facades\Deprecator;
 use CraftCms\Yii2Adapter\Validation\LegacyElementRules;
@@ -17,7 +23,8 @@ final readonly class ValidateMixin
             Deprecator::log($this::class . '->hasErrors', 'Calling `->hasErrors` is deprecated. Use `->errors()->has($attribute)` or `->errors()->isNotEmpty()` instead.');
 
             /**
-             * @var \CraftCms\Cms\Element\Element|\CraftCms\Cms\Field\Field|\CraftCms\Cms\Field\Field $this
+             * @var Element|Field|Field $this
+             *
              * @phpstan-ignore-next-line
              */
             return is_null($attribute)
@@ -32,7 +39,8 @@ final readonly class ValidateMixin
             Deprecator::log($this::class . '->getErrors', 'Calling `->getErrors` is deprecated. Use `->errors()->get($attribute)` or `->errors()->getMessages()` instead.');
 
             /**
-             * @var \CraftCms\Cms\Element\Element|\CraftCms\Cms\Field\Field $this
+             * @var Element|Field $this
+             *
              * @phpstan-ignore-next-line
              */
             return is_null($attribute)
@@ -47,7 +55,8 @@ final readonly class ValidateMixin
             Deprecator::log($this::class . '->addErrors', 'Calling `->addErrors` is deprecated. Use `->errors()->add($attribute, $message)` instead.');
 
             /**
-             * @var \CraftCms\Cms\Element\Element|\CraftCms\Cms\Field\Field $this
+             * @var Element|Field $this
+             *
              * @phpstan-ignore-next-line
              */
             $this->errors()->add($attribute, $error);
@@ -60,7 +69,8 @@ final readonly class ValidateMixin
             Deprecator::log($this::class . '->addError', 'Calling `->addError` is deprecated. Use `->errors()->add($attribute, $message)` instead.');
 
             /**
-             * @var \CraftCms\Cms\Asset\Data\Volume|\CraftCms\Cms\Dashboard\Widgets\Widget|\CraftCms\Cms\Element\Element|\CraftCms\Cms\Field\Field|\CraftCms\Cms\FieldLayout\FieldLayoutComponent|\CraftCms\Cms\Filesystem\Filesystems\Filesystem $this
+             * @var Volume|Widget|Element|Field|FieldLayoutComponent|Filesystem $this
+             *
              * @phpstan-ignore-next-line
              */
             $this->errors()->add($attribute, $error);
@@ -74,7 +84,8 @@ final readonly class ValidateMixin
 
             if ($attribute === null) {
                 /**
-                 * @var \CraftCms\Cms\Element\Element|\CraftCms\Cms\Field\Field $this
+                 * @var Element|Field $this
+                 *
                  * @phpstan-ignore-next-line
                  */
                 foreach ($this->errors()->getMessages() as $key => $messages) {
@@ -86,7 +97,8 @@ final readonly class ValidateMixin
             }
 
             /**
-             * @var \CraftCms\Cms\Element\Element|\CraftCms\Cms\Field\Field $this
+             * @var Element|Field $this
+             *
              * @phpstan-ignore-next-line
              */
             $this->errors()->forget($attribute);
@@ -99,7 +111,8 @@ final readonly class ValidateMixin
             Deprecator::log($this::class . '->getFirstError', 'Calling `->getFirstError` is deprecated. Use `->getFirstErrors()` instead.');
 
             /**
-             * @var \CraftCms\Cms\Element\Element|\CraftCms\Cms\Field\Field $this
+             * @var Element|Field $this
+             *
              * @phpstan-ignore-next-line
              */
             return Arr::get($this->getFirstErrors(), $attribute);
@@ -112,7 +125,8 @@ final readonly class ValidateMixin
             Deprecator::log($this::class . '->getAttributeLabel', 'Calling `->getAttributeLabel` is deprecated. Use `->attributeLabels()` instead.');
 
             /**
-             * @var \CraftCms\Cms\Asset\Data\Volume|\CraftCms\Cms\Dashboard\Widgets\Widget|\CraftCms\Cms\Element\Element|\CraftCms\Cms\Field\Field|\CraftCms\Cms\FieldLayout\FieldLayoutComponent|\CraftCms\Cms\Filesystem\Filesystems\Filesystem $this
+             * @var Volume|Widget|Element|Field|FieldLayoutComponent|Filesystem $this
+             *
              * @phpstan-ignore-next-line
              */
             return $this->attributeLabels()[$attribute] ?? $attribute;
